update 0.6.5.0-Alpha6-Patch8 and Hollow
-由于代码修正众多，特定提交两次
-本次代码改动了以下方面：
1.迭代了部分素材
2.添加破碎的天狗BGM,并且优化天狗的策略
3.修复一些分数的扣除错误
4.灯火现已支持小数扣减
5.一些底层的修改变动
6.矮人国王额外召唤重甲豺狼和冰雪魔像
7.Yog二阶段额外召唤两个拳头

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Nyctophobia.java
@@ -66,7 +66,7 @@ public class Nyctophobia extends Buff implements Hero.Doom {
             }
             spend(90f);
         }  else if(hero.lanterfire < 0){
-            hero.damage( (Dungeon.depth/5)*(Math.abs(hero.lanterfire)/2)+10, this );
+            hero.damage( (int)((Dungeon.depth/5)*(Math.abs(hero.lanterfire)/2)+10), this );
             spend(24f);
             GLog.w( Messages.get(this, "desc6") );
         }
